🐞 Detects virtual machine interfaces
Previously, when hoping from the IP address to the interface associated
with that IP, NetboxClientRuby would always return a `DCIM::Interface`
instance.

```ruby
# would always return a DCIM::Interface
NetboxClientRuby.ip_addresses.first.interface
```

This is wrong, because if the IP address is assigned to a virtual
machine, then we would expected to receive a
`Virtualization::Interface` instead.

This commit fixes #24.