Support Faraday 2
BREAKING CHANGE: To add support for Faraday 2, the dependency on
"faraday_middleware" must be dropped. Netbox Client Ruby uses
"faraday_middleware" to process request/response with JSON
encoding/decoding middleware for Faraday 0.x and 1.x. In Faraday 2, this
middleware is built-in. Netbox Client Ruby will still attempt to require
"faraday_middleware" to support Faraday 0.x and 1.x and will raise an
error if it is not installed. Users of Netbox Client Ruby must install
"faraday_middleware" separately to continue using with Faraday 0.x or
1.x.